Fix some typos in line rasterization methods
Had to debug this when running Rasters.  It's actually kind of slow on a FeatureCollection - will have to figure out (a) why and (b) how to optimize this.

The end goal is to create a kind of `DatashaderAxis`, which can be plotted into, and extracts geometries from all those plots, then datashades them.

It would be nice to have some kind of "online" rasterization interface as well, would make the workflow a lot cleaner.  I can see a function `rasterize!`