Updated error message when attempting to read a file without the appropriate backend. (#465)
* Updated error message when attempting to read a file without the appropriate backend.

The previous error message for reading rasters without an imported backend was confusing. The new message informs callers of the change and suggests the appropriate backend to use.

* Update sources.jl

* Updated error message when attempting to read a file without the appropriate backend.

The previous error message for reading rasters without an imported backend was confusing. The new message informs callers of the change and suggests the appropriate backend to use.

* Updated error message when attempting to read a file without the appropriate backend.

The previous error message for reading rasters without an imported backend was confusing. The new message informs callers of the change and suggests the appropriate backend to use.

The error now reads, "`Rasters.jl` requires backends to be loaded externally as of version 0.8. Run `import $packagename` to read $filename"