Throw method error or backend exception correctly (#480)
* Updated error message when attempting to read a file without the appropriate backend.

The previous error message for reading rasters without an imported backend was confusing. The new message informs callers of the change and suggests the appropriate backend to use.

* Update sources.jl

* Updated error message when attempting to read a file without the appropriate backend.

The previous error message for reading rasters without an imported backend was confusing. The new message informs callers of the change and suggests the appropriate backend to use.

* Updated error message when attempting to read a file without the appropriate backend.

The previous error message for reading rasters without an imported backend was confusing. The new message informs callers of the change and suggests the appropriate backend to use.

The error now reads, "`Rasters.jl` requires backends to be loaded externally as of version 0.8. Run `import $packagename` to read $filename"

* Added warning to docs regarding extensions

* Added warning to docs regarding extensions

* Added warning to docs regarding extensions

* resample and warp now throw MethodError instead of BackendException after loading ArchGDAL

* check extentsion

* remove duplicate method stubs

* include Test

* only test extensions on 1.9

* fix for 1.8

---------

Co-authored-by: JoshuaBillson <61667893+JoshuaBillson@users.noreply.github.com>
Co-authored-by: JoshuaBillson <jmbillson@outlook.com>