Add ZarrDataset extension (#640)
* Add ZarrDataset extension

This is a first draft for opening zarr data with ZarrDatasets in Rasters

* Remove pre 1.9 check for extensions

* Fix dirty rebase

* Fix Zarrsource package link

* Add ZarrDatasets as test dep

* Start implementing write for Zarrsource

* Treat Zarr folder as single file in RasterStack

Otherwise the RasterStack code tries to load all zarr subfiles separately.

* Add check for locus

This always assumes a Center locus when the bounds are not explicitely hit.
We might want to check for the actual center, but this might become problematic for time bounds.

* Replace all by mapreduce

* Add some tests using the ESDL as testfile

* Fix typo

* Add more tests

This copied some of the ncdatasets tests and did them on a Zarr array.

* Update registry in Documenter action

This is necessary, because otherwise the  documentation build fails with strange pkg errors.

* Move _writevar! to Rasters

I am trying to make the _writevar! function only depend on CDM functions to make it backend agnostic.
This needs certain functions to make the NCD specific parts inside functions which can be specialized on the source type.

* Move NCDataset specific error handling into NCDatasetsExt

* Remove write function

* Dispatch _check_allowed_type on sourctrait

* Remove show macro calls