Implement a `bylayer` keyword for `zonal`
if false, `f` gets an iterable of named tuples.  if true, `f` gets an iterable of values, and we use `maplayers` outside that.

Still an open question what the default should be.  I would like to have it be `true` to underline the nature of RasterStacks as being arrays of namedtuples as well, but `false` would be less breaking.